requirements.txt
streamlit==1.45.1
duckduckgo_search==8.0.4
faiss==1.7.4
numpy==1.25.2
transformers==4.31.0
torch==2.0.1
yfinance==0.2.25
google-genai==1.20.0
plotly==5.15.0
pandas==2.3.0
requests==2.31.0
python-dotenv==1.1.0
